Rename schemas/pending_announcements -> pending_announcement
