+++ imports/client/components/NotificationCenter.tsx
@@ -26,7 +26,7 @@ import { AnnouncementType } from '../../lib/schemas/announcement';
 import { ChatNotificationType } from '../../lib/schemas/chat_notification';
 import { GuessType } from '../../lib/schemas/guess';
 import { HuntType } from '../../lib/schemas/hunt';
-import { PendingAnnouncementType } from '../../lib/schemas/pending_announcements';
+import { PendingAnnouncementType } from '../../lib/schemas/pending_announcement';
 import { PuzzleType } from '../../lib/schemas/puzzles';
 import { guessURL } from '../../model-helpers';
 import { requestDiscordCredential } from '../discord';
