+++ imports/lib/models/pending_announcements.ts
@@ -1,8 +1,8 @@
-import PendingAnnouncementsSchema, { PendingAnnouncementType } from '../schemas/pending_announcements';
+import PendingAnnouncementSchema, { PendingAnnouncementType } from '../schemas/pending_announcement';
 import Base from './base';
 
 const PendingAnnouncements = new Base<PendingAnnouncementType>('pending_announcements');
-PendingAnnouncements.attachSchema(PendingAnnouncementsSchema);
+PendingAnnouncements.attachSchema(PendingAnnouncementSchema);
 PendingAnnouncements.publish(function (q) {
   // It's sufficient to use the user property for filtering here; we
   // don't need to pay attention to the hunt ID
